Diabetes Check
Diabetes is a medical condition that is characterized by elevated level of sugar content in the blood. This occurs when the body's ability to create insulin is affected. However,a person's insulin level is not the only parameter to be checked to see if he/she has diabetes. There are other parameters using which we can predict if a person  is diabetes positive or not. By subjecting the dataset that contains all the parameters to a machine learning algorithm, we can predict if a person has diabetes or not.